pyproject.toml
[project]
name = "spam-bots-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
